reset the iTunes controller when iTunes gets terminated
Listen for the NSWorkspaceDidTerminateApplicationNotification
notification and reset the controller.

The problem is that the iTunes playerInfo notification gets sent before
the terminate notification which causes a hang in
updateControllerStatus because iTunes is no longer
accessible through the used processID.
However the playerInfo notifications userInfo
dictionary consists only of one entry which is the
"Player State".
Now the "Player State" and the dictionary size get
checked to determine if iTunes is terminating or not.